requirements.txt
python-telegram-bot==21.*
fastapi==0.115.*
uvicorn==0.30.*
google-auth==2.*
google-auth-oauthlib==1.*
google-api-python-client==2.*
beautifulsoup4==4.*
openai==1.*
httpx==0.27.*
python-dotenv==1.*
